src/requirements.txt
gelfclient==0.0.7
spacy==2.3.5
python_docx==0.8.11
absl-py==1.4.0
alembic==1.8.1
bayoo_docx==0.2.19
beautifulsoup4==4.12.2
docx==0.2.4
Flask==2.2.5
Flask_Cors==3.0.10
Flask_Migrate==4.0.0
Flask_RESTful==0.3.9
flask_sqlalchemy
gelfclient==0.0.7
joblib==1.2.0
marshmallow==3.19.0
nltk==3.8.1
pandas==2.0.3
pypandoc_binary==1.11
python-dotenv==1.0.0
Requests==2.31.0
scikit_learn==0.24.2
tqdm==4.64.1
gunicorn==21.2.0
pymysql==1.1.1
aspose_words